Update Project : Re-structured Base Page
